Replace the ring-servlet monkey patch with configuator
Manipulating global state should be minimalized for moving to a
component base workflow. Remove the alter-var-root monkey patch.
Replace it with a decorator around the http response object that
will set the status message based on the headers.

Theoretically this could be slower since it has a check everytime
a header is set. However, the jit might handle it. An alternative
is to decorate the outputstream and writer and only perform the
check on .flush.