Beginning of duct based project w/ repl based dev
This commit is larger then I was hoping as it touches several parts.

1. Configuration through the command line was removed. It was ignored
in migrations and setup. Configuration is now read on startup based on
env and config.clj.  It remains a global and will probably be one of the
last pieces of global state to be isolated. (This probably could have
been a subpatch.)

2. The jetty server is now a component. Architecture is moving towards
the duct library/template. This is the outermost point in the system
and needed to be the first to be a component. As additional parts of
the system are made into components, it will be possible to isolate
and test them cleanly.

3. A repl based reloaded workflow has been introduced. This allows
starting, stopping, refreshing files, and testing to occur from the
repl.