Move yeller into a component and protocolize api
Several places were using error reporting through yeller. This
patch threads through a record to those places.

A protocol is used to allow replacement by an implementation that
ignores it for developement and testing. This requires creating a yeller
wrapper for the protocol. Since yeller has a client creation function,
it has a lifecycle. This was wrapped into a component. A system can then
inject the proper handler.

The error wrapping middleware is applied only in the production system,
allowing wrap-stacktrace to show a complete error page in development.

Finally, some of the tests were reduced to testing a smaller part of the
system as the api they were calling changed.