Replace korma with yesql
Normally reading the database config once is great, but korma
does it globally. So lets replace it. I choose yesql as it lets
sql stay sql.

Now the database functions will read the config each time. This is
unfortunate, but allows a dev and test config to be used in the
same jvm.  The next step will be to make the db into a component
to allow defining a thread pool around the db config.