Make the database a component
One of the most used global state was the database config. This
commit changes almost everything to use it as a parameter. A
threadpool is added to handle the component-ization.

This does require a backwards incompatible change to the way
database configuation can occur. From now on they should be a
jdbc url.

The tests see significant changes to account for the new apis.
Additionally the fixures change to allow keeping one connection
to the database per test, instead of reading the config each line.
Tests also become more restricted to using just the ui if possible.
All together this reduces test runtime by 10 seconds on my machine.

The clojars admin namespace contains a dynamic var for its database
connection.  The nrepl server uses some middleware to populate that
var for connections.

The one off tasks (migrations,setup) are changed to use the new apis.
They just use the database string from the configuration. They do not
try to setup a threadpool or system.