WIP experimental commit for parallel tests
This patch adds parallelizing the tests. Unfortunatly with-redefs and
multithreading does not work. This commit removes all with-redefs that
could interfere with the tests, and they can run parallelize.
Unfortunatly the config change is still done with a with-redef and needs
to be removed. This commit also starts the process by removing some uses
of the global config.